+++ reproc/src/pipe.posix.c
@@ -14,8 +14,8 @@
 
 const int PIPE_INVALID = -1;
 
-const int PIPE_EVENT_IN = POLLIN;
-const int PIPE_EVENT_OUT = POLLOUT;
+const short PIPE_EVENT_IN = POLLIN;
+const short PIPE_EVENT_OUT = POLLOUT;
 
 int pipe_init(int *read, int *write)
 {
@@ -110,7 +110,7 @@ int pipe_poll(pipe_event_source *sources, size_t num_sources, int timeout)
 
   for (size_t i = 0; i < num_sources; i++) {
     pollfds[i].fd = sources[i].pipe;
-    pollfds[i].events = (short) sources[i].interests;
+    pollfds[i].events = sources[i].interests;
   }
 
   r = poll(pollfds, (nfds_t) num_sources, timeout);
